requirements.txt
flask==2.0.1
flask-sqlalchemy==2.5.1
flask-mail==0.9.1
python-dotenv==0.19.0
openai==0.27.0
apscheduler==3.9.1
gunicorn==20.1.0
werkzeug==2.0.1
jinja2==3.0.1
itsdangerous==2.0.1
click==8.0.1
requests==2.26.0
python-dateutil==2.8.2
sqlalchemy==1.4.23
apscheduler
flask
flask-mail
flask-sqlalchemy
openai
python-dotenv
sqlalchemy
psycopg2-binary
